[+] Add iterated/evolutionary prisoner dilemma features
With Generation & Evolution it's possible to make PD run over several rounds (generations) and make the different strategies fight for survival.
The population of a generation is static but the population of a strategy changes over time.
Given X or all the possible strategies, have multiple prisoners for each strategy.
The strategies that perform well as a collective they gain in population over different generations.
The strategies that perform badly as a collective, decay in population over generations until they die off.