[+] Iterate further on data structure around Strategy vs Prisoner
Tip: Who's meant to have which piece of information?
Description, identity...that belongs to the strategy itself.
The prisoner is the person, not their behavior, that's something they inherit from the strategy itself.

The strategies are the ones meant to have unique ID (although, due to Rust, it's a weak identity)
They are also meant to tell about being nice as so on.

It would be nice to have the behavior associated to the data itself, but, once again, Rust does not seem to support this in a straight forward way.
Meaning, this is a nice iteration for the code in order to learn and so on, certain thing. Simply not happening at this very same moment.